hr-video-analyzer/requirements.txt
fastapi>=0.70.0
uvicorn[standard]>=0.15.0
opencv-python>=4.5.3.56
python-multipart>=0.0.5
numpy>=1.19.0
textblob>=0.17.1
vaderSentiment>=3.3.2
nltk>=3.8
openai-whisper>=20231117
ffmpeg-python>=0.2.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
